Add gitignore, Add packages.json, Add react packages and move TodoItem into his own file
