+++ src/index.html
@@ -9,10 +9,6 @@
 
     <div id="root"></div> <!-- create a "parent" div especially for react app otherwise everything in it will be overwritten -->
 
-    <script src="https://unpkg.com/react@16.13.1/umd/react.development.js"></script>
-    <script src="https://unpkg.com/react-dom@16.13.1/umd/react-dom.development.js"></script>
-
-
     <script src="./App.js"></script>
 </body>
 </html>
